+++ index.html
@@ -6,6 +6,7 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <title>ProjectManager</title>
     <link rel="stylesheet" href="app.css" />
+    <script>var exports = {};</script>
     <script src="dist/app.js" defer></script>
   </head>
   <body>
